requirements.txt
pandas==2.1.0
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
SQLAlchemy==2.0.23
schedule==1.2.0
pytz==2023.3
streamlit==1.31.0
plotly==5.18.0 